리뷰어의 권고 사항에 따라 두 가지 개선 사항을 모두 반영했습니다.
DictPreservingQueueHandler 중복 제거: 순환 참조를 방지하고 코드 재사용성을 높이기 위해 core/loggers/async_handler.py라는 새로운 모듈로 분리하고, 필요한 곳에서 임포트하여 사용하도록 수정했습니다.
중앙 집중식 로거 종료 처리: core/logger.py에 shutdown_logging() 함수를 추가하여 애플리케이션이나 테스트 종료 시 남아있는 백그라운드 큐와 파일 핸들을 안전하게 정리할 수 있도록 조치했습니다.

main.py 파일의 애플리케이션 종료 지점(finally 블록)에 shutdown_logging() 호출을 추가하여 백그라운드 큐와 파일 핸들러가 안전하게 정리되도록 코드를 수정했습니다.
