+++ core/loggers/async_handler.py
@@ -1,0 +1,9 @@
+from logging.handlers import QueueHandler
+
+class DictPreservingQueueHandler(QueueHandler):
+    def prepare(self, record):
+        original_msg = record.msg
+        record = super().prepare(record)
+        if isinstance(original_msg, dict):
+            record.msg = original_msg
+        return record
